+++ src/byte_buffer.hpp
@@ -11,29 +11,29 @@ public:
     ByteBuffer();
     ByteBuffer(const ByteBuffer & copy);
     ByteBuffer(const char * str);
-    ByteBuffer(const char * str, int length);
+    ByteBuffer(const char * str, size_t length);
     ~ByteBuffer() {}
     ByteBuffer& operator= (const ByteBuffer& other);
 
-    const char & at(int index) const {
+    const char & at(size_t index) const {
         return _buffer.at(index);
     }
-    char & at(int index) {
+    char & at(size_t index) {
         return _buffer.at(index);
     }
 
     static ByteBuffer format(const char *format, ...) __attribute__ ((format (printf, 1, 2)));
 
-    int length() const {
+    size_t length() const {
         return _buffer.length() - 1;
     }
-    void resize(int new_length) {
+    void resize(size_t new_length) {
         _buffer.resize(new_length + 1);
         _buffer.at(length()) = 0;
     }
     void append(const ByteBuffer &other);
     void append(const char *str);
-    void append(const char *str, int length);
+    void append(const char *str, size_t length);
 
     void append_uint32le(uint32_t value) {
         // 00000000 00000000 00000000 xxxxxxxx
@@ -55,7 +55,7 @@ public:
         _buffer.at(length()) = value;
         _buffer.append(0);
     }
-    uint32_t read_uint32le(int index) const {
+    uint32_t read_uint32le(size_t index) const {
         if (index < 0 || index + 4 > length())
             panic("bounds check");
         char *buf = _buffer.raw() + index;
@@ -64,23 +64,23 @@ public:
                (((uint32_t)buf[2]) << 16) |
                (((uint32_t)buf[3]) << 24) ;
     }
-    uint8_t read_uint8(int index) const {
+    uint8_t read_uint8(size_t index) const {
         if (index < 0 || index >= length())
             panic("bounds check");
         char *buf = _buffer.raw() + index;
         return (uint8_t)*buf;
     }
-    void append_fill(int count, char value) {
-        int index = length();
+    void append_fill(size_t count, char value) {
+        size_t index = length();
         resize(index + count);
         char *buf = _buffer.raw() + index;
         memset(buf, value, count);
     }
 
-    int index_of_rev(char c) const;
-    int index_of_rev(char c, int start) const;
+    size_t index_of_rev(char c) const;
+    size_t index_of_rev(char c, size_t start) const;
 
-    ByteBuffer substring(int start, int end) const;
+    ByteBuffer substring(size_t start, size_t end) const;
 
     char *raw() {
         return _buffer.raw();
@@ -99,7 +99,7 @@ public:
     uint32_t hash() const {
         // FNV 32-bit hash
         uint32_t h = 2166136261;
-        for (int i = 0; i < _buffer.length(); i += 1) {
+        for (size_t i = 0; i < _buffer.length(); i += 1) {
             h = h ^ ((uint8_t)_buffer.at(i));
             h = h * 16777619;
         }
