+++ src/gui.cpp
@@ -72,7 +72,7 @@ Gui::Gui(SDL_Window *window, ResourceBundle *resource_bundle,
 }
 
 Gui::~Gui() {
-    for (int i = 0; i < _widget_list.length(); i += 1) {
+    for (size_t i = 0; i < _widget_list.length(); i += 1) {
         Widget *widget = _widget_list.at(i);
         destroy_widget(widget);
     }
@@ -196,7 +196,7 @@ void Gui::exec() {
 
         glClear(GL_COLOR_BUFFER_BIT|GL_DEPTH_BUFFER_BIT|GL_STENCIL_BUFFER_BIT);
 
-        for (int i = 0; i < _widget_list.length(); i += 1) {
+        for (size_t i = 0; i < _widget_list.length(); i += 1) {
             Widget *widget = _widget_list.at(i);
             if (widget->_is_visible) {
                 widget->draw(widget, _projection);
@@ -358,7 +358,7 @@ void Gui::on_mouse_move(const MouseEvent *event) {
     if (_mouse_over_widget != NULL)
         panic("expected _mouse_over_widget NULL");
 
-    for (int i = 0; i < _widget_list.length(); i += 1) {
+    for (size_t i = 0; i < _widget_list.length(); i += 1) {
         Widget *widget = _widget_list.at(i);
         if (try_mouse_move_event_on_widget(widget, event))
             return;
