+++ src/hash_map.hpp
@@ -11,14 +11,14 @@ public:
     HashMap() {
         init_capacity(32);
     }
-    HashMap(int capacity) {
+    HashMap(size_t capacity) {
         init_capacity(capacity);
     }
     ~HashMap() {
         destroy(_entries, _capacity);
     }
 
-    int size() const {
+    size_t size() const {
         return _size;
     }
 
@@ -29,10 +29,10 @@ public:
         // if we get too full (80%), double the capacity
         if (_size * 5 >= _capacity * 4) {
             Entry *old_entries = _entries;
-            int old_capacity = _capacity;
+            size_t old_capacity = _capacity;
             init_capacity(_capacity * 2);
             // dump all of the old elements into the new table
-            for (int i = 0; i < old_capacity; i += 1) {
+            for (size_t i = 0; i < old_capacity; i += 1) {
                 Entry *old_entry = &old_entries[i];
                 if (old_entry->used)
                     internal_put(old_entry->key, old_entry->value);
@@ -59,9 +59,9 @@ public:
 
     void remove(const K &key) {
         _modification_count += 1;
-        int start_index = key_to_index(key);
-        for (int roll_over = 0; roll_over <= _max_distance_from_start_index; roll_over += 1) {
-            int index = (start_index + roll_over) % _capacity;
+        size_t start_index = key_to_index(key);
+        for (size_t roll_over = 0; roll_over <= _max_distance_from_start_index; roll_over += 1) {
+            size_t index = (start_index + roll_over) % _capacity;
             Entry *entry = &_entries[index];
 
             if (!entry->used)
@@ -71,7 +71,7 @@ public:
                 continue;
 
             for (; roll_over < _capacity; roll_over += 1) {
-                int next_index = (start_index + roll_over + 1) % _capacity;
+                size_t next_index = (start_index + roll_over + 1) % _capacity;
                 Entry *next_entry = &_entries[next_index];
                 if (!next_entry->used || next_entry->distance_from_start_index == 0) {
                     entry->used = false;
@@ -110,9 +110,9 @@ public:
     private:
         HashMap * _table;
         // how many items have we returned
-        int _count = 0;
+        size_t _count = 0;
         // iterator through the entry array
-        int _index = 0;
+        size_t _index = 0;
         // used to detect concurrent modification
         uint32_t _inital_modification_count;
         Iterator(HashMap * table) :
@@ -129,34 +129,34 @@ private:
 
     struct Entry {
         bool used;
-        int distance_from_start_index;
+        size_t distance_from_start_index;
         K key;
         V value;
     };
 
     Entry *_entries;
-    int _capacity;
-    int _size;
-    int _max_distance_from_start_index;
+    size_t _capacity;
+    size_t _size;
+    size_t _max_distance_from_start_index;
     // this is used to detect bugs where a hashtable is edited while an iterator is running.
     uint32_t _modification_count = 0;
 
-    void init_capacity(int capacity) {
+    void init_capacity(size_t capacity) {
         _capacity = capacity;
         _entries = allocate<Entry>(_capacity);
         _size = 0;
         _max_distance_from_start_index = 0;
-        for (int i = 0; i < _capacity; i += 1) {
+        for (size_t i = 0; i < _capacity; i += 1) {
             _entries[i].used = false;
         }
     }
 
     void internal_put(K key, V value) {
-        int start_index = key_to_index(key);
-        for (int roll_over = 0, distance_from_start_index = 0;
+        size_t start_index = key_to_index(key);
+        for (size_t roll_over = 0, distance_from_start_index = 0;
                 roll_over < _capacity; roll_over += 1, distance_from_start_index += 1)
         {
-            int index = (start_index + roll_over) % _capacity;
+            size_t index = (start_index + roll_over) % _capacity;
             Entry *entry = &_entries[index];
 
             if (entry->used && entry->key != key) {
@@ -199,9 +199,9 @@ private:
 
 
     Entry *internal_get(const K &key) const {
-        int start_index = key_to_index(key);
-        for (int roll_over = 0; roll_over <= _max_distance_from_start_index; roll_over += 1) {
-            int index = (start_index + roll_over) % _capacity;
+        size_t start_index = key_to_index(key);
+        for (size_t roll_over = 0; roll_over <= _max_distance_from_start_index; roll_over += 1) {
+            size_t index = (start_index + roll_over) % _capacity;
             Entry *entry = &_entries[index];
 
             if (!entry->used)
@@ -213,8 +213,8 @@ private:
         return NULL;
     }
 
-    int key_to_index(const K &key) const {
-        return (int)(HashFunction(key) % ((uint32_t)_capacity));
+    size_t key_to_index(const K &key) const {
+        return (size_t)(HashFunction(key) % ((uint32_t)_capacity));
     }
 };
