+++ src/text_widget.cpp
@@ -273,8 +273,8 @@ void TextWidget::update_selection_model() {
 }
 
 void TextWidget::set_selection(int start, int end) {
-    _cursor_start = clamp(0, start, _label.text().length());
-    _cursor_end = clamp(0, end, _label.text().length());
+    _cursor_start = clamp(0, start, (int)_label.text().length());
+    _cursor_end = clamp(0, end, (int)_label.text().length());
     scroll_cursor_into_view();
 }
 
@@ -304,13 +304,13 @@ void TextWidget::on_text_input(const TextInputEvent *event) {
 }
 
 void TextWidget::replace_text(int start, int end, const String &text, int cursor_modifier) {
-    start = clamp(0, start, _label.text().length());
-    end = clamp(0, end, _label.text().length());
+    start = clamp(0, start, (int)_label.text().length());
+    end = clamp(0, end, (int)_label.text().length());
 
     _label.replace_text(start, end, text);
 
-    _cursor_start = clamp(0, start + cursor_modifier, _label.text().length());
-    _cursor_end = clamp(0, start + cursor_modifier, _label.text().length());
+    _cursor_start = clamp(0, start + cursor_modifier, (int)_label.text().length());
+    _cursor_end = clamp(0, start + cursor_modifier, (int)_label.text().length());
 
     _label.update();
     scroll_cursor_into_view();
@@ -441,9 +441,9 @@ int TextWidget::advance_word(int dir) {
 
 int TextWidget::advance_word_from_index(int start_index, int dir) {
     int init_advance = (dir > 0) ? 0 : 1;
-    int new_cursor = clamp(0, start_index - init_advance, _label.text().length() - 1);
+    int new_cursor = clamp(0, start_index - init_advance, (int)_label.text().length() - 1);
     bool found_non_whitespace = false;
-    while (new_cursor >= 0 && new_cursor < _label.text().length()) {
+    while (new_cursor >= 0 && (size_t)new_cursor < _label.text().length()) {
         uint32_t c = _label.text().at(new_cursor);
         if (String::is_whitespace(c)) {
             if (found_non_whitespace)
